requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pyngrok==7.1.6
groq==0.9.0
requests==2.33.0